requirements.txt
asgiref==3.2.10
awsebcli==3.19.2
botocore==1.19.25
cachetools==4.1.1
cement==2.8.2
certifi==2020.6.20
chardet==3.0.4
checksum==1.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-bootstrap3==14.1.0
django-config==0.1.4
django-ical==1.7.1
django-paypal==1.0.0
django-recurrence==1.10.3
future==0.16.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
httplib2==0.18.1
icalendar==4.0.7
idna==2.10
jmespath==0.10.0
oauthlib==3.1.0
pathspec==0.5.9
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pypiwin32==223
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
pywin32==300
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
semantic-version==2.5.0
six==1.15.0
sqlparse==0.3.1
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.25.10
wcwidth==0.1.9
whitenoise==5.2.0
